requirements.txt
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
google-api-core==2.29.0
google-api-python-client==2.190.0
google-api-python-client-stubs==1.32.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-auth-stubs==0.3.0
googleapis-common-protos==1.72.0
grpc-stubs==1.53.0.6
grpcio==1.78.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
msal==1.34.0
oauthlib==3.3.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.6
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
types-httplib2==0.31.2.20260125
types-requests==2.32.4.20260107
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
